fix package specified C type error

+++ godablooms/dablooms.go
@@ -20,20 +20,20 @@ type ScalingBloom struct {
 	cfilter *C.scaling_bloom_t
 }
 
-func NewScalingBloom(capacity C.uint, errorRate C.double, filename string) *ScalingBloom {
+func NewScalingBloom(capacity uint, errorRate float64, filename string) *ScalingBloom {
 	cFilename := C.CString(filename)
 	defer C.free(unsafe.Pointer(cFilename))
 	sb := &ScalingBloom{
-		cfilter: C.new_scaling_bloom(capacity, errorRate, cFilename),
+		cfilter: C.new_scaling_bloom(C.uint(capacity), C.double(errorRate), cFilename),
 	}
 	return sb
 }
 
-func NewScalingBloomFromFile(capacity C.uint, errorRate C.double, filename string) *ScalingBloom {
+func NewScalingBloomFromFile(capacity uint, errorRate float64, filename string) *ScalingBloom {
 	cFilename := C.CString(filename)
 	defer C.free(unsafe.Pointer(cFilename))
 	sb := &ScalingBloom{
-		cfilter: C.new_scaling_bloom_from_file(capacity, errorRate, cFilename),
+		cfilter: C.new_scaling_bloom_from_file(C.uint(capacity), C.double(errorRate), cFilename),
 	}
 	return sb
 }
@@ -50,24 +50,24 @@ func (sb *ScalingBloom) Check(key []byte) bool {
 	return C.scaling_bloom_check(sb.cfilter, cKey, C.size_t(len(key))) == 1
 }
 
-func (sb *ScalingBloom) Add(key []byte, id C.uint64_t) bool {
+func (sb *ScalingBloom) Add(key []byte, id uint64) bool {
 	cKey := (*C.char)(unsafe.Pointer(&key[0]))
-	return C.scaling_bloom_add(sb.cfilter, cKey, C.size_t(len(key)), id) == 1
+	return C.scaling_bloom_add(sb.cfilter, cKey, C.size_t(len(key)), C.uint64_t(id)) == 1
 }
 
-func (sb *ScalingBloom) Remove(key []byte, id C.uint64_t) bool {
+func (sb *ScalingBloom) Remove(key []byte, id uint64) bool {
 	cKey := (*C.char)(unsafe.Pointer(&key[0]))
-	return C.scaling_bloom_remove(sb.cfilter, cKey, C.size_t(len(key)), id) == 1
+	return C.scaling_bloom_remove(sb.cfilter, cKey, C.size_t(len(key)), C.uint64_t(id)) == 1
 }
 
 func (sb *ScalingBloom) Flush() bool {
 	return C.scaling_bloom_flush(sb.cfilter) == 1
 }
 
-func (sb *ScalingBloom) MemSeqNum() C.uint64_t {
-	return C.scaling_bloom_mem_seqnum(sb.cfilter)
+func (sb *ScalingBloom) MemSeqNum() uint64 {
+	return uint64(C.scaling_bloom_mem_seqnum(sb.cfilter))
 }
 
-func (sb *ScalingBloom) DiskSeqNum() C.uint64_t {
-	return C.scaling_bloom_disk_seqnum(sb.cfilter)
+func (sb *ScalingBloom) DiskSeqNum() uint64 {
+	return uint64(C.scaling_bloom_disk_seqnum(sb.cfilter))
 }
